+++ src/commonMain/kotlin/teksturepako/pakku/api/actions/fetch/Fetch.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.channels.produce
 import teksturepako.pakku.api.actions.errors.*
 import teksturepako.pakku.api.data.ConfigFile
 import teksturepako.pakku.api.data.LockFile
-import teksturepako.pakku.api.http.Http
+import teksturepako.pakku.api.http.requestByteArray
 import teksturepako.pakku.api.platforms.Provider
 import teksturepako.pakku.api.projects.ProjectFile
 import java.nio.file.Path
@@ -55,12 +55,12 @@ suspend fun List<ProjectFile>.fetch(
                     totalBytes += projectFile.size.toLong()
                     val prevBytes: AtomicLong = atomic(0L)
 
-                    val bytes = Http().requestByteArray(projectFile.url!!) { bytesSentTotal, _ ->
+                    val bytes = requestByteArray(projectFile.url!!) { bytesSentTotal, _ ->
                         completedBytes.getAndAdd(bytesSentTotal - prevBytes.value)
 
                         onProgress(completedBytes.value, totalBytes.value)
                         prevBytes.getAndSet(bytesSentTotal)
-                    }
+                    }.get()
 
                     if (bytes == null)
                     {